+++ src/main/java/pl/plajer/villagedefense3/Main.java
@@ -246,10 +246,7 @@ public class Main extends JavaPlugin {
               Bukkit.getConsoleSender().sendMessage(ChatColor.YELLOW + "Current version: " + ChatColor.RED + currentVersion + ChatColor.YELLOW + " Latest version: " + ChatColor.GREEN + latestVersion);
             }
           }
-        } catch (Exception ex) {
-          Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "[VillageDefense] An error occured while checking for update!");
-          Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "Please check internet connection or check for update via WWW site directly!");
-          Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "WWW site https://www.spigotmc.org/resources/41869");
+        } catch (Exception ignored) {
         }
       }